JvmInfo response
